Remove deprecated files/methods
