--- examples/modelspecs-best-psfg2.csv
@@ -1,13 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-gausspx,mgsersic4:1,nser,nser=0.5,moments,gaussian:2,T
-mg4exppx,mgsersic4:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg4devepx,mgsersic4:1,nser,nser=4,guessexp2dev:mg4exppx,gaussian:2,T
-mg4serbpx,mgsersic4:1,,,best,gaussian:2,T
-mg4devexppx,mgsersic4:2,"nser","nser=4,1",mg4devepx;mg4exppx,gaussian:2,T
-mg4serserbpx,mgsersic4:2,,,best,gaussian:2,T
-mg4wserpx,gaussian:4+rscale:1,cenx;ceny;nser;rscale;sigma_x;sigma_y;rho,rho=inherit;rscale=modify,mg4serbpx,gaussian:2,T
-mg4wrserpx,gaussian:4+rscale:1,nser;sigma_x;sigma_y,rho=inherit;rscale=modify,mg4wserpx,gaussian:2,T
-mg4aserbpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify,mg4wrserpx,gaussian:2,T
-mg4x2rpx,gaussian:8+rscale:2,nser;sigma_x;sigma_y,rho=inherit;rscale=modify,mg4serserbpx,gaussian:2,T
-mg4x2px,gaussian:8+rscale:2,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4serserbpx,gaussian:2,T
-mg4x2bpx,gaussian:8+rscale:2,nser;rscale,rho=inherit;rscale=modify,mg4x2rpx,gaussian:2,T
