--- examples/modelspecs-mg-no-rs-psfg2.csv
@@ -1,40 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-gausspx,mgsersic8:1,nser,nser=0.5,moments,gaussian:2,T
-mg8sermpx,mgsersic8:1,,nser=1,moments,gaussian:2,T
-mg8serggpx,mgsersic8:1,,,guessgauss2ser:gausspx,gaussian:2,T
-mg8expgpx,mgsersic8:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg8expmpx,mgsersic8:1,nser,nser=1,moments,gaussian:2,T
-mg8devepx,mgsersic8:1,nser,nser=4,guessexp2dev:mg8expmpx,gaussian:2,T
-mg8devgpx,mgsersic8:1,nser,nser=4,guessgauss2dev:gausspx,gaussian:2,T
-mg8devmpx,mgsersic8:1,nser,nser=4,moments,gaussian:2,T
-mg8serbpx,mgsersic8:1,,,best,gaussian:2,T
-mg8serbedpx,mgsersic8:1,,,best:mg8expgpx;mg8devepx,gaussian:2,T
-mg8cmodelpx,mgsersic8:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg8devepx;mg8expgpx",gaussian:2,T
-mg8devexppx,mgsersic8:2,nser,"nser=4,1",mg8devepx;mg8expgpx,gaussian:2,T
-mg8devexptpx,mgsersic8:2,"cenx;ceny;flux;nser","nser=4,1",mg8devepx;mg8expgpx,gaussian:2,T
-mg8devexp2px,mgsersic8:2,nser,"nser=4,1",mg8devexptpx,gaussian:2,T
-mg8devexpcpx,mgsersic8:2,nser,"nser=4,1",mg8cmodelpx,gaussian:2,T
-mg8serserbpx,mgsersic8:2,,,best,gaussian:2,T
-mg4sermpx,mgsersic4:1,,nser=1,moments,gaussian:2,T
-mg4expgpx,mgsersic4:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg4expmpx,mgsersic4:1,nser,nser=1,moments,gaussian:2,T
-mg4devepx,mgsersic4:1,nser,nser=4,guessexp2dev:mg8expgpx,gaussian:2,T
-mg4devgpx,mgsersic4:1,nser,nser=4,guessgauss2dev:gausspx,gaussian:2,T
-mg4devmpx,mgsersic4:1,nser,nser=4,moments,gaussian:2,T
-mg4serggpx,mgsersic4:1,,,guessgauss2ser:gausspx,gaussian:2,T
-mg4serbpx,mgsersic4:1,,,best,gaussian:2,T
-mg4serbedpx,mgsersic4:1,,,best:mg4expgpx;mg4devepx,gaussian:2,T
-mg4cmodelpx,mgsersic4:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg4devepx;mg4expgpx",gaussian:2,T
-mg4devexppx,mgsersic4:2,nser,"nser=4,1",mg4devepx;mg4expgpx,gaussian:2,T
-mg4devexptpx,mgsersic4:2,"cenx;ceny;flux;nser","nser=4,1",mg4devepx;mg4expgpx,gaussian:2,T
-mg4devexp2px,mgsersic4:2,nser,"nser=4,1",mg4devexptpx,gaussian:2,T
-mg4devexpcpx,mgsersic4:2,nser,"nser=4,1",mg4cmodelpx,gaussian:2,T
-mg4serserbpx,mgsersic4:2,,,best,gaussian:2,T
-mg4faexppx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4expmpx,gaussian:2,T
-mg4fadevpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4devmpx,gaussian:2,T
-mg4faserpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4serbpx,gaussian:2,T
-mg8faexppx,gaussian:8+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8expmpx,gaussian:2,T
-mg8fadevpx,gaussian:8+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8devmpx,gaussian:2,T
-mg8faserpx,gaussian:8+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8serbpx,gaussian:2,T
-mg4fax2px,gaussian:8+rscale:2,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4serserbpx,gaussian:2,T
-mg8fax2px,gaussian:16+rscale:2,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8serserbpx,gaussian:2,T
