--- examples/modelspecs-mg-psfg2.csv
@@ -1,45 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-gausspx,mgsersic8:1,nser,nser=0.5,moments,gaussian:2,T
-mg8serggpx,mgsersic8:1,,,guessgauss2ser:gausspx,gaussian:2,T
-mg8exppx,mgsersic8:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg8devepx,mgsersic8:1,nser,nser=4,guessexp2dev:mg8exppx,gaussian:2,T
-mg8devgpx,mgsersic8:1,nser,nser=4,guessgauss2dev:gausspx,gaussian:2,T
-mg8serbpx,mgsersic8:1,,,best,gaussian:2,T
-mg8serbedpx,mgsersic8:1,,,best:mg8exppx;mg8devepx,gaussian:2,T
-mg8cmodelpx,mgsersic8:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg8devepx;mg8exppx",gaussian:2,T
-mg8devexppx,mgsersic8:2,nser,"nser=4,1",mg8devepx;mg8exppx,gaussian:2,T
-mg8devexptpx,mgsersic8:2,"cenx;ceny;flux;nser","nser=4,1",mg8devepx;mg8exppx,gaussian:2,T
-mg8devexp2px,mgsersic8:2,nser,"nser=4,1",mg8devexptpx,gaussian:2,T
-mg8devexpcpx,mgsersic8:2,nser,"nser=4,1",mg8cmodelpx,gaussian:2,T
-mg8serserbpx,mgsersic8:2,,,best,gaussian:2,T
-mg4exppx,mgsersic4:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg4devepx,mgsersic4:1,nser,nser=4,guessexp2dev:mg8exppx,gaussian:2,T
-mg4devgpx,mgsersic4:1,nser,nser=4,guessgauss2dev:gausspx,gaussian:2,T
-mg4serggpx,mgsersic4:1,,,guessgauss2ser:gausspx,gaussian:2,T
-mg4serbpx,mgsersic4:1,,,best,gaussian:2,T
-mg4serbedpx,mgsersic4:1,,,best:mg4exppx;mg4devepx,gaussian:2,T
-mg4cmodelpx,mgsersic4:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg4devepx;mg4exppx",gaussian:2,T
-mg4devexppx,mgsersic4:2,nser,"nser=4,1",mg4devepx;mg4exppx,gaussian:2,T
-mg4devexptpx,mgsersic4:2,"cenx;ceny;flux;nser","nser=4,1",mg4devepx;mg4exppx,gaussian:2,T
-mg4devexp2px,mgsersic4:2,nser,"nser=4,1",mg4devexptpx,gaussian:2,T
-mg4devexpcpx,mgsersic4:2,nser,"nser=4,1",mg4cmodelpx,gaussian:2,T
-mg4serserbpx,mgsersic4:2,,,best,gaussian:2,T
-mg4wexppx,gaussian:4+rscale:1,cenx;ceny;nser;rscale;sigma_x;sigma_y;rho,rho=inherit;rscale=modify,mg4exppx,gaussian:2,T
-mg4wrexppx,gaussian:4+rscale:1,nser;re,rho=inherit;rscale=modify,mg4exppx,gaussian:2,T
-mg4aexppx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4exppx,gaussian:2,T
-mg4aexpbpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify,mg4wrexppx,gaussian:2,T
-mg4wdevpx,gaussian:4+rscale:1,cenx;ceny;nser;rscale;sigma_x;sigma_y;rho,rho=inherit;rscale=modify,mg4devepx,gaussian:2,T
-mg4wrdevpx,gaussian:4+rscale:1,nser;re,rho=inherit;rscale=modify,mg4devepx,gaussian:2,T
-mg4adevpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4devepx,gaussian:2,T
-mg4adevbpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify,mg4wrdevpx,gaussian:2,T
-mg4wserpx,gaussian:4+rscale:1,cenx;ceny;nser;rscale;sigma_x;sigma_y;rho,rho=inherit;rscale=modify,mg4serbpx,gaussian:2,T
-mg4wrserpx,gaussian:4+rscale:1,nser;re,rho=inherit;rscale=modify,mg4serbpx,gaussian:2,T
-mg4aserpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4serbpx,gaussian:2,T
-mg4aserbpx,gaussian:4+rscale:1,nser;rscale,rho=inherit;rscale=modify,mg4wrserpx,gaussian:2,T
-mg8wserpx,gaussian:8+rscale:1,cenx;ceny;nser;rscale;sigma_x;sigma_y;rho,rho=inherit;rscale=modify,mg8serbpx,gaussian:2,T
-mg8wrserpx,gaussian:8+rscale:1,nser;re,rho=inherit;rscale=modify,mg8serbpx,gaussian:2,T
-mg8aserpx,gaussian:8+rscale:1,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8serbpx,gaussian:2,T
-mg8aserbpx,gaussian:8+rscale:1,nser;rscale,rho=inherit;rscale=modify,mg8wrserpx,gaussian:2,T
-mg4x2rpx,gaussian:8+rscale:2,nser;re,rho=inherit;rscale=modify,mg4serserbpx,gaussian:2,T
-mg4x2px,gaussian:8+rscale:2,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg4serserbpx,gaussian:2,T
-mg4x2bpx,gaussian:8+rscale:2,nser;rscale,rho=inherit;rscale=modify,mg4x2rpx,gaussian:2,T
