--- examples/modelspecs-mg8-base-cmodel-psfg2.csv
@@ -1,16 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-gausspx,mgsersic8:1,nser,nser=0.5,moments,gaussian:2,T
-mg8sermpx,mgsersic8:1,,nser=1,moments,gaussian:2,T
-mg8serggpx,mgsersic8:1,,,guessgauss2ser:gausspx,gaussian:2,T
-mg8expgpx,mgsersic8:1,nser,nser=1,guessgauss2exp:gausspx,gaussian:2,T
-mg8expmpx,mgsersic8:1,nser,nser=1,moments,gaussian:2,T
-mg8devepx,mgsersic8:1,nser,nser=4,guessexp2dev:mg8expmpx,gaussian:2,T
-mg8devgpx,mgsersic8:1,nser,nser=4,guessgauss2dev:gausspx,gaussian:2,T
-mg8devmpx,mgsersic8:1,nser,nser=4,moments,gaussian:2,T
-mg8serbpx,mgsersic8:1,,,best,gaussian:2,T
-mg8serbedpx,mgsersic8:1,,,best:mg8expgpx;mg8devepx,gaussian:2,T
-mg8cmodelpx,mgsersic8:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg8devepx;mg8expgpx",gaussian:2,T
-mg8devexppx,mgsersic8:2,nser,"nser=4,1",mg8devepx;mg8expgpx,gaussian:2,T
-mg8devexptpx,mgsersic8:2,"cenx;ceny;flux;nser","nser=4,1",mg8devepx;mg8expgpx,gaussian:2,T
-mg8devexp2px,mgsersic8:2,nser,"nser=4,1",mg8devexptpx,gaussian:2,T
-mg8devexpcpx,mgsersic8:2,nser,"nser=4,1",mg8cmodelpx,gaussian:2,T
