--- examples/modelspecs-mg8-base-linear-psfg2.csv
@@ -1,7 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-mg8sermpx,mgsersic8:1,,nser=1,moments,gaussian:2,T
-mg8serapx,gaussian:8+rscale:1,cenx;ceny;nser;sigma_x;sigma_y;rscale;rho,rho=inherit;rscale=modify,mg8sermpx,gaussian:2,T
-mg8devexptpx,mgsersic8:2,"cenx;ceny;flux;nser","nser=4,1",mg8sermpx;mg8sermpx,gaussian:2,T
-mg8devexp2px,mgsersic8:2,nser,"nser=4,1",mg8devexptpx,gaussian:2,T
-mg8serserbpx,mgsersic8:2,,,best,gaussian:2,T
-mg8serserapx,gaussian:16+rscale:2,cenx;ceny;nser;sigma_x;sigma_y;rscale;rho,rho=inherit;rscale=modify,mg8serserbpx,gaussian:2,T
