--- examples/modelspecs-mg8-sermpx-psfg2.csv
@@ -1,2 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-mg8sermpx,mgsersic8:1,,nser=1,moments,gaussian:2,T
