--- examples/modelspecs-plot-final-psfg2.csv
@@ -1,11 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-gausspx,mgsersic8:1,nser,nser=0.5,moments,gaussian:2,T
-mg8expgx,mgsersic8:1,nser,nser=1,gausspx,gaussian:2,T
-mg8devepx,mgsersic8:1,nser,nser=4,mg8exppx,gaussian:2,T
-mg8serbpx,mgsersic8:1,,,best,gaussian:2,T
-mg8cmodelpx,mgsersic8:2,"cenx;ceny;nser;sigma_x;sigma_y;rho","nser=4,1","mg8dev2px;mg8exppx",gaussian:2,T
-mg8devexp2px,mgsersic8:2,nser,"nser=4,1",mg8dev2px;mg8exppx,gaussian:2,T
-mg8serserbpx,mgsersic8:2,,"nser=4,1",mg8dev2px;mg8exppx,gaussian:2,T
-serbpx,sersic:1,,,mg8serbpx,gaussian:2,T
-serb,sersic:1,,,mg8serbpx,gaussian:2,F
-mg8fax2px,gaussian:16+rscale:2,nser;rscale,rho=inherit;rscale=modify;rscale=1,mg8serserbpx,gaussian:2,T
