--- examples/modelspecs-serbpx-psfg2.csv
@@ -1,3 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-mg8serbpx,mgsersic8:1,,,moments,gaussian:2,T
-serbpx,sersic:1,,,mg8serbpx,gaussian:2,T
