--- examples/modelspecs-sersic-psfg2.csv
@@ -1,3 +1,0 @@
-name,model,fixedparams,initparams,inittype,psfmodel,psfpixel
-sermpx,sersic:1,,nser=1,moments,gaussian:2,T
-serb,sersic:1,,,sermpx,gaussian:2,F
